Allow setting stride equal to width of the clip
Current check did not allow stride to be set equal to width of the clip,
once it was set to a larger value.
In one of the AdaptivePlaybackTests, stride is set to a value larger
than current width before decoding the header and then set to value
equal to width as buffers are allocated as per the new width.
Decoder was expecting the stride to be larger value, thus resulting in
buffer size errors.

Bug: 124531341
Test: cts-tradefed run commandAndExit cts-dev -m CtsMediaTestCases \
 -t android.media.cts.AdaptivePlaybackTest#testMpeg2_adaptiveDrc

Change-Id: I056aa943ab4281d268926a670bef7e9503227cb5